Addedd move functionality for all pieces except for the PAWN.
In this commit Wertzuz and I have updated several piece functions to add functionality where the pieces to eat the same colored pieces.
For this we have come up with a function called findindmoved which can be used in all findmoves function except for the pawn function. in the next commit this will be implemented.